refactor: 배포시 발생한 에러 수정을 위한 폴더 구조 변경
