+++ pages/folder/index.tsx
@@ -1,16 +1,16 @@
 import styles from "./folderPage.module.css";
 import { KeyboardEvent, useEffect, useRef, useState } from "react";
-import { getUserLinks } from "../api/folder";
-import { BASE_URL, USERS_ENDPOINT } from "../api/services/config";
+import { getUserLinks } from "../../api/folder";
+import { BASE_URL, USERS_ENDPOINT } from "../../api/services/config";
 import { FolderName, LinkInfo } from "@/types/types";
-import { ALL_LINK_NAME, OPTION_ICONS } from "./constant";
-import AddLinkInput from "./components/addLinkInput/AddLinkInput";
+import { ALL_LINK_NAME, OPTION_ICONS } from "../../constants/folderConstant";
+import AddLinkInput from "../../components/addLinkInput/AddLinkInput";
 import SearchBar from "@/components/searchBar/SearchBar";
 import Card from "@/components/card/Card";
-import EmptyPage from "./components/emptyPage/EmptyPage";
-import OptionButton from "./components/optionButton/OptionButton";
+import EmptyPage from "../../components/emptyPage/EmptyPage";
+import OptionButton from "../../components/optionButton/OptionButton";
 import FloatingButton from "@/components/floatingButton/FloatingButton";
-import SortButton from "./components/sortButton/SortButton";
+import SortButton from "../../components/sortButton/SortButton";
 import addIcon from "@/public/icons/add.svg";
 import addPrimaryIcon from "@/public/icons/addPrimaryColor.svg";
 import useModal from "@/hooks/useModal";
